Closes #180, #181, #183

+++ src/specialfuns.jl
@@ -178,7 +178,7 @@ end
 
 
 # Multidimensional gamma / partial gamma function
-function lpgamma(p::Int64, a::Float64)
+function lpgamma(p::Int, a::Float64)
     res::Float64 = p * (p - 1.0) / 4.0 * log(pi)
     for ii in 1:p
         res += lgamma(a + (1.0 - ii) / 2.0)
